requirements.txt
langchain
openai
python-dotenv
pytest
azure-identity